blog_app/requirements/local.txt
-r ./base.txt

# Django
# ------------------------------------------------------------------------------
django-debug-toolbar==2.2

# Developer Tools
# ------------------------------------------------------------------------------
ipdb==0.13.2
ipython==7.14.0
flake8==3.8.2

# Testing
# ------------------------------------------------------------------------------
mock==4.0.2
django-nose==1.4.6
nose-progressive==1.5.2
coverage==5.1
